Switch from Uint16Array to Uint8Array for more efficient encoding/decoding

+++ src/html/parseHTML.ts
@@ -102,7 +102,7 @@ const appendContentToNode = async (parent: YetiRootNode | YetiElementNode | Open
  * The parser works by first lexing the input HTML string into a stream of tokens using the lexHTML function,
  * and then processing each token to build up the node tree.
  */
-export const parseHTML = async (htmlStringChars: Uint16Array, dynamicValues: unknown[]): Promise<YetiRootNode> => {
+export const parseHTML = async (htmlStringChars: Uint8Array, dynamicValues: unknown[]): Promise<YetiRootNode> => {
   const rootNode: YetiRootNode = { type: YETI_NODE_TYPE.ROOT, children: [] };
 
   let openParentStack: Array<YetiElementNode | OpenComponentNode> = [];
